+++ src/Entity/DPDPriceRule.php
@@ -14,6 +14,11 @@ use Invertus\dpdBaltics\Repository\DPDZoneRepository;
 use Invertus\dpdBaltics\Repository\PriceRuleRepository;
 use Invertus\dpdBaltics\Repository\ZoneRepository;
 
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+
 /**
  * Class DPDPriceRule
  */
@@ -257,10 +262,14 @@ class DPDPriceRule extends ObjectModel
         /** @var DPDBaltics $module */
         $module = Module::getInstanceByName('dpdbaltics');
 
-        /** @var ZoneRepository $zoneRepo */
-        $zoneRepo = $module->getModuleContainer()->get('invertus.dpdbaltics.repository.zone_repository');
+        /** @var \Invertus\dpdBaltics\Verification\IsAddressInZone $isAddressInZone */
+        $isAddressInZone = $module->getModuleContainer('invertus.dpdbaltics.verification.is_address_in_zone');
 
-        return $zoneRepo->findZoneInRangeByAddress($address);
+        /** @var DPDZoneRepository $repo */
+        $repo = $module->getModuleContainer()->get('invertus.dpdbaltics.repository.dpdzone_repository');
+        $zonesIds = $repo->getZonesIdsByPriceRule($this->id);
+
+        return $isAddressInZone->verify($address, $zonesIds);
     }
 
     public function isApplicableByPrice($price, $idCurrency)