+++ src/Infrastructure/Bootstrap/Install/Installer.php
@@ -1,8 +1,21 @@
 <?php
-
+/**
+ * NOTICE OF LICENSE
+ *
+ * @author    INVERTUS, UAB www.invertus.eu <support@invertus.eu>
+ * @copyright Copyright (c) permanent, INVERTUS, UAB
+ * @license   Addons PrestaShop license limitation
+ * @see       /LICENSE
+ *
+ * International Registered Trademark & Property of INVERTUS, UAB
+ */
 namespace Invertus\dpdBaltics\Infrastructure\Bootstrap\Install;
 
 use Invertus\dpdBaltics\Infrastructure\Bootstrap\Exception\CouldNotInstallModule;
+
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
 
 class Installer implements InstallerInterface
 {