+++ CHANGELOG.md
@@ -169,4 +169,9 @@
 ## [3.2.20]
 - One page checkout compatibility improvements
 - Multiple payment transaction and total price fix
+
+## [3.2.21]
+- Fixed label printing error and 0 goods price displaying
+- Improved price calculation
+- Fixed issues with carrier shown as "Free" when it is not needed to display
 - Carriers error handling improvements